Replace invalid build-tags and use correct dup2 syscall.
Previously, we was using 'unix' during compilation as a build-tag,
in the syscall files, which is not a valid GOOS—subsequently causing
error during compilation on a number of architectures. This also
unbreaks 32-bit compilation.

In 'syscall_unix.go' 3 arguments was being passed to the 'dup2'
syscall (when 2 is only needed) also breaking compilation, which
is now fixed.